// platform/app/src/telepacs/api/generated/model/decisionEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TelePACS Backend API
 * Versioned API contract for the independent TelePACS backend.
 * OpenAPI spec version: 0.1.0
 */

/**
 * * `APPROVE` - APPROVE
 * * `REQUEST_CHANGES` - REQUEST_CHANGES
 * * `REJECT` - REJECT
 */
export type DecisionEnum = typeof DecisionEnum[keyof typeof DecisionEnum];


export const DecisionEnum = {
  APPROVE: 'APPROVE',
  REQUEST_CHANGES: 'REQUEST_CHANGES',
  REJECT: 'REJECT',
} as const;