// platform/app/src/telepacs/api/generated/model/doctorCredentialReviewState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TelePACS Backend API
 * Versioned API contract for the independent TelePACS backend.
 * OpenAPI spec version: 0.1.0
 */

/**
 * * `DRAFT` - Draft
 * * `SUBMITTED` - Submitted
 * * `APPROVED` - Approved
 * * `CHANGES_REQUESTED` - Changes requested
 * * `REJECTED` - Rejected
 * * `REVOKED` - Revoked
 * * `EXPIRED` - Expired
 */
export type DoctorCredentialReviewState = typeof DoctorCredentialReviewState[keyof typeof DoctorCredentialReviewState];


export const DoctorCredentialReviewState = {
  DRAFT: 'DRAFT',
  SUBMITTED: 'SUBMITTED',
  APPROVED: 'APPROVED',
  CHANGES_REQUESTED: 'CHANGES_REQUESTED',
  REJECTED: 'REJECTED',
  REVOKED: 'REVOKED',
  EXPIRED: 'EXPIRED',
} as const;